refactor: extract useSortableListEditor hook and SnapshotCard component
- SortableListEditor: move all state/handlers into useSortableListEditor,
  component reduced to sensors + JSX only (~200 lines → hook + ~170 lines)
- SnapshotPanel: extract per-snapshot card JSX into SnapshotCard component
  (including its own rename input ref and focus effect),
  panel reduced from 398 to ~260 lines
